inbox: own the Person/Organization identity specs
Moves the capability supplier from plugin-space to plugin-inbox. Whoever creates
the objects owns the rule for deciding when two of them are the same, and these
are materialised from mail — plugin-space renders the review but has no business
knowing which types have identities. plugin-inbox is also enabled by default, so
the Duplicates tab is available wherever mail sync is.

The capability seam is unchanged; other plugins still contribute rules for their
own types.
