+++ packages/plugins/plugin-inbox/src/capabilities/index.ts
@@ -9,6 +9,7 @@ export const AppGraphBuilder = Capability.lazy('AppGraphBuilder', () => import('
 export const SkillDefinition = Capability.lazy('SkillDefinition', () => import('./skill-definition'));
 export const CreateObject = Capability.lazy('CreateObject', () => import('./create-object'));
 export const Connector = Capability.lazy('Connector', () => import('./connector'));
+export const IdentitySpecs = Capability.lazy('IdentitySpecs', () => import('./identity-specs'));
 export const NavigationTargetResolver = Capability.lazy(
   'NavigationTargetResolver',
   () => import('./navigation-target-resolver'),
