+++ packages/plugins/plugin-space/src/SpacePlugin.ts
@@ -29,7 +29,6 @@ import {
   AppGraphBuilder,
   CreateObject,
   IdentityCreated,
-  IdentitySpecs,
   NavigationHandler,
   NavigationTargetResolver,
   OperationHandler,
@@ -138,11 +137,6 @@ export const SpacePlugin = Plugin.define<SpacePluginOptions>(meta).pipe(
     activatesOn: ClientEvents.IdentityCreated,
     firesAfterActivation: [SpaceEvents.PersonalSpaceReady],
     activate: IdentityCreated,
-  }),
-  Plugin.addModule({
-    id: `${meta.profile.key}/module/identity-specs`,
-    activatesOn: AppActivationEvents.SetupSchema,
-    activate: IdentitySpecs,
   }),
   Plugin.addModule({
     activatesOn: ActivationEvent.allOf(
